requirements.txt
# ============================================
# Legal RAG Chatbot - Dependencies
# ============================================

# --- Core API ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0

# --- Data Processing ---
pandas>=2.2.0
pyarrow>=15.0.0

# --- Search & Retrieval ---
whoosh>=2.7.4
faiss-cpu>=1.7.4
numpy>=1.26.0

# --- LLM Integration ---
ollama>=0.4.0
httpx>=0.27.0

# --- Text Processing ---
underthesea>=6.8.0
regex>=2024.0.0

# --- Evaluation ---
rouge-score>=0.1.2
scikit-learn>=1.4.0

# --- Logging ---
loguru>=0.7.0
